+++ src/cli.js
@@ -1,9 +1,10 @@
 #! /usr/bin/env node
 'use strict';
 
-const util = require('util');
+const castArray = require('lodash.castarray');
 const program = require('commander');
 const tplinkCrypto = require('tplink-smarthome-crypto');
+const util = require('util');
 
 const { Client, ResponseError } = require('./');
 
@@ -12,17 +13,19 @@ let client;
 
 const outputError = function (err) {
   if (err instanceof ResponseError) {
+    console.log('Response Error:');
     console.log(err.response);
   } else {
+    console.error('Error:');
     console.error(err);
   }
 };
 
-let search = function (sysInfo, timeout, params) {
+const search = function (sysInfo, breakoutChildren, timeout, params) {
   try {
     console.log('Searching...');
 
-    let commandParams = Object.assign({}, { discoveryInterval: 2000, discoveryTimeout: timeout }, params); // {discoveryInterval: 2000, discoveryTimeout: timeout, ...params};
+    const commandParams = Object.assign({}, { discoveryInterval: 2000, discoveryTimeout: timeout, breakoutChildren }, params); // {discoveryInterval: 2000, discoveryTimeout: timeout, ...params};
     console.log(`startDiscovery(${util.inspect(commandParams)})`);
     client.startDiscovery(commandParams)
       .on('device-new', (device) => {
@@ -47,7 +50,7 @@ const send = async function (host, port, payload) {
   }
 };
 
-let sendCommand = async function (host, port, payload) {
+const sendCommand = async function (host, port, childId, payload) {
   try {
     console.log(`Sending to ${host}:${port}...`);
     const device = await client.getDevice({ host, port });
@@ -59,8 +62,10 @@ let sendCommand = async function (host, port, payload) {
   }
 };
 
-let sendCommandDynamic = async function (host, port, command, commandParams = []) {
+const sendCommandDynamic = async function (host, port, command, commandParams = [], childId = null) {
   try {
+    console.log(`Sending ${command} command to ${host}:${port} ${childId ? 'childId: ' + childId : ''}...`);
+    const device = await client.getDevice({ host, port, childId });
     const results = await device[command](...commandParams);
     console.log('response:');
     console.dir(results, { colors: program.color === 'on', depth: 10 });
@@ -116,7 +121,7 @@ const setParamTypes = function setParamTypes (params, types) {
     return castArray(params).map((el, i) => {
       switch (types[i]) {
         case 'number':
-        return +el;
+          return +el;
         case 'boolean':
           return (el === 'true' || el === '1');
       }
@@ -128,7 +133,7 @@ const setParamTypes = function setParamTypes (params, types) {
 
 program
   .option('-D, --debug', 'turn on debug level logging', () => { logLevel = 'debug'; })
-  .option('-t, --timeout <ms>', 'timeout (ms)', toInt, 5000)
+  .option('-t, --timeout <ms>', 'timeout (ms)', toInt, 10000)
   .option('-u, --udp', 'send via UDP')
   .option('-c, --color [on]', 'output will be styled with ANSI color codes', 'on');
 
@@ -136,13 +141,14 @@ program
   .command('search [params]')
   .description('Search for devices')
   .option('-s, --sysinfo', 'output sysInfo')
+  .option('-b, --breakout-children', 'output children (multi-outlet plugs)', true)
   .action(function (params, options) {
     client = setupClient();
     if (params) {
       console.dir(params);
       params = JSON.parse(params);
     }
-    search(options.sysinfo, program.timeout, params);
+    search(options.sysinfo, options.breakoutChildren || false, program.timeout, params);
   });
 
 program
@@ -179,27 +185,37 @@ program
     blink(hostOnly, port, times, rate);
   });
 
-[ 'getSysInfo', 'getInfo', 'setAlias', 'getModel',
+[ { fnName: 'getSysInfo', supportsChildId: true },
+  { fnName: 'getInfo', supportsChildId: true },
+  { fnName: 'setAlias', supportsChildId: true },
+  { fnName: 'getModel', supportsChildId: true },
+  { fnName: 'setPowerState', paramTypes: ['boolean'], supportsChildId: true },
   { fnName: 'setLocation', paramTypes: ['number', 'number'] },
   { fnName: 'reboot', paramTypes: ['number'] },
   { fnName: 'reset', paramTypes: ['number'] }
 ].forEach((command) => {
   let commandName;
   let paramTypes;
+  let supportsChildId = false;
   if (command.fnName) {
     commandName = command.fnName;
     paramTypes = command.paramTypes;
+    supportsChildId = command.supportsChildId;
   } else {
     commandName = command;
   }
-  program
+  let cmd = program
     .command(`${commandName} <host> [params]`)
     .description(`Send ${commandName} to device (using Device#${commandName})`)
-    .option('-t, --timeout [timeout]', 'timeout (ms)', toInt, 5000)
-    .action(function (host, params, options) {
-      client = setupClient();
+    .option('-t, --timeout [timeout]', 'timeout (ms)', toInt, 10000);
+  if (supportsChildId) {
+    cmd = cmd.option('-c, --childId [childId]', 'childId');
+  }
+  cmd.action(function (host, params, options) {
+    client = setupClient();
     const [hostOnly, port] = host.split(':');
-    });
+    sendCommandDynamic(hostOnly, port, commandName, setParamTypes(params, paramTypes), options.childId);
+  });
 });
 
 program